Simple version of the Extended Query Machine
It turns out that Postgresql has an optimization that allows the client to send multiple requests to the backend without expecting a response from each individual one.
This is done to save some TCP round-trips (but it's pretty annoying since when you do want those responses, you have to pay an extra round-trip for each one.)

In any case, this adds the `Flush` message which forces the backend to produce the expected response. It's not optimal, but it allows the current dispatcher to still work in a request/response style (though it requires knowledge of the `Flush` messages).