Adds named prepared statement caching.
This implements named prepared statement chaching, similar to how finagle-mysql handles them.

Named prepared statements are part of a connection's state. Thus, to be able to reuse an already prepared statement,
a client would have to make sure to always issue requests on the same connection (the one onto which the statement was prepared).

This complicates connection pooling since we'd like to be able to use them interchangebly.

Thus, a cache is introduced in the `ClientDispatcher`. This cache holds the response of previously prepared statements which avoids re-preparing statements on the same connection
as well as allows releasing the connection for re-use.