fix: update pg_hbase.conf to allow local
Something changed in GH actions that started making the `HandshakeSpec` fail.
It turns out that it now needs to allow `local` connections which is strange since those should be for connections over the unix domain socket.
Perhaps something changed in Docker that makes connections look like they're from the domain socket now.

This shouldn't work since `local` is for unix domain socket connections which isn't what we're using, but we'll see.